move ExpQueue unit tests to utils
